Switch from /api/crowbar/upgrade to /api/upgrade.
The overall upgrade status API at /api/upgrade gives more details about
the situation in the backend and is implemented also in crowbar-init.
This is important during admin server upgrade as crowbar is shut down
during the upgrade and can't provide status information.

This change switches from /api/crowbar/upgrade to /api/upgrade in the
context of admin server upgrade. The old service for accessing crowbar
upgrade status is not removed.