upgrade: Extracted polling for current step's status
The polling/waiting for current step to end needs to be used in
multiple pages so to avoid code duplication and simplify testing
this functionality was extracted into a separate service.

Admin server upgrade step was updated to use the new service.
Other places where it needs to be used will be done separately.