upgrade: Fix ongoing upgrade status handling
When page was reloaded during admin server upgrade the upgrade
button would get disabled and start spinning but polling for status
would not be (re)started.

This change extracts the common "wait for upgrade step to end" part
into a separate function and uses it in both cases (new and ongoing
upgrade).