Asynchronous prepare step & status downtime (#85)
* upgrade: Extracted polling for current step's status

The polling/waiting for current step to end needs to be used in
multiple pages so to avoid code duplication and simplify testing
this functionality was extracted into a separate service.

Admin server upgrade step was updated to use the new service.
Other places where it needs to be used will be done separately.

* upgrade: Updated Landing page to poll for prepare status

The prepare operation is done in background so the UI needs to
poll for status of this step to keep user on the landing page until
prepare is finished.

* upgrade: Allow temporary downtime when polling for status

In some cases (e.g. admin server upgrade) the status API becomes
unavailable temporarily due to node reboot or service restart.
Those downtimes should not be treated as errors if they don't last
too long.

This change introduces additional parameter to waitForStepToEnd()
function to enable caller to specify how much downtime is acceptable.
First usage for this is admin server upgrade step mentioned above.

* upgrade: Fix ongoing upgrade status handling

When page was reloaded during admin server upgrade the upgrade
button would get disabled and start spinning but polling for status
would not be (re)started.

This change extracts the common "wait for upgrade step to end" part
into a separate function and uses it in both cases (new and ongoing
upgrade).

* upgrade: moved string literals to constants

Some strings like step names and statuses are used in multiple places
so it is better to have them defined as constants. With this approach
the compiler will help spot possible typos which will be harder to spot
with literal string.