upgrade: Restore last step on refresh or reopen
When browser window is closed and reopened or refreshed during upgrade flow,
the user should be presented with the most appropriate view of the flow.
The UI will show the page respective for `current_step` as returned by the
status API and/or state stored in browser.

Same mechanism is used to prevent user from jumping to any page in the flow
by manually entering the page URL into the address bar.