upgrade: Incrementally complete items in checklist
In some cases, checklist is used to show status of long running
process with multiple steps. One global `completed` flag is not enough
in such cases as the view will be updated after all steps are completed.

This change introduces an alternative way to control completion of check
items. Client should not set global `completed` attribute in the template
but set per-check `completed` field in the checks structure passed to the
directive.