upgrade: Split services/backup into two pages
As those two steps are not really related, the UI was modified so that
they are handled on two separate pages. The logic is way less complicated
after this change as there's no need to 'enforce' the sequential execution
of those steps inside the controller or show the status of those 'sub-steps'.